split lf paradox/nexwatch into own files
see changelog.md for cli changes!
