+++ client/Makefile
@@ -99,6 +99,8 @@ CMDSRCS = 	crapto1/crapto1.c\
 			cmdlfhitag.c \
 			cmdlfio.c \
 			cmdlfindala.c \
+			cmdlfnexwatch.c \
+			cmdlfparadox.c \
 			cmdlfpcf7931.c\
 			cmdlfpresco.c\
 			cmdlfpyramid.c\
