+++ client/cmdlfnexwatch.h
@@ -1,0 +1,14 @@
+//-----------------------------------------------------------------------------
+//
+// This code is licensed to you under the terms of the GNU GPL, version 2 or,
+// at your option, any later version. See the LICENSE.txt file for the text of
+// the license.
+//-----------------------------------------------------------------------------
+// Low frequency Honeywell NexWatch tag commands
+//-----------------------------------------------------------------------------
+#ifndef CMDLFNEXWATCH_H__
+#define CMDLFNEXWATCH_H__
+extern int CmdLFNexWatch(const char *Cmd);
+extern int CmdFSKdemodParadox(const char *Cmd);
+extern int CmdParadoxRead(const char *Cmd);
+#endif
