+++ client/cmdlfparadox.h
@@ -1,0 +1,14 @@
+//-----------------------------------------------------------------------------
+//
+// This code is licensed to you under the terms of the GNU GPL, version 2 or,
+// at your option, any later version. See the LICENSE.txt file for the text of
+// the license.
+//-----------------------------------------------------------------------------
+// Low frequency Paradox tag commands
+//-----------------------------------------------------------------------------
+#ifndef CMDLFPARADOX_H__
+#define CMDLFPARADOX_H__
+extern int CmdLFParadox(const char *Cmd);
+extern int CmdFSKdemodParadox(const char *Cmd);
+extern int CmdParadoxRead(const char *Cmd);
+#endif
